mxtx-dgramtunneld.c: fix multible "clients" bind same port 40001
When SO_REUSEADDR is (apparently) set in DGRAM (udp) socket,
it allows other processes to bind the same socket (and start
receiving messages to it) (my '90's learned rules w/ tcp does not
apply here).

Fixed that by not setting SO_REUSEADDR -- now process will get
EADDRINUSE (not the ECONNREFUSED that was accidentally put there)
if udp port is already in use -- and dgramtunneld tries the next
one (40002 (and so on))

Now mxtx-mosh connections to multiple mxtx destinations work at
the same time (or first again also after second exited).