mxtx-socksproxy: opt to use tcp 127.1:port. recognize (fail) http[s]/tls
With :port as 1st arg tcp socket on localhost:port is bound and listened
instead of the unix socket (to where one usually needs ld_preload to
access). In addition to some temporary use, it is harder for some
sandboxed browsers to "inject" that ld_preload.

The user (peer) who connects is not checked when this option is used...

In case of (accidental) http/https proxy configuration the first
byte of the request may be also 'G', 'C' or 0x16 (22):

'G' (probably GET ...)
'C' (probably CONNECT ...)
 22 (probably TLS handshake, Client hello) (*)

(*) https://tls13.xargs.org/#client-hello/annotated (there as of 202510)

These are now recognized and hint that user has "misconfigured" their
client is printed.

These (and one another thing) were tested with something like...

  curl -v -x socks5://127.0.0.1 http://127.1:8888/
  curl -v -x socks5h://127.0.0.1 http://srvr:8888/
  curl -v -x http://127.0.0.1 http://srvr:8888/
  curl -v -x https://127.0.0.1 http://srvr:8888/
  curl -v -x https://127.0.0.1:1080 http://srvr:8888/
  strace curl -x https://127.0.0.1:1080 http://srvr:8888/
  strace -e trace=network curl -x https://127.1:1080 http://srvr:8888/
  curl --trace - -x https://127.1:1080 http://srvr:8888/